+++ tensorrt_llm/_utils.py
@@ -46,11 +46,6 @@ from tensorrt_llm.bindings import DataType, GptJsonConfig, LayerType
 from tensorrt_llm.bindings.BuildInfo import ENABLE_MULTI_DEVICE
 from tensorrt_llm.logger import logger
 
-try:
-    import ray
-except ImportError:
-    ray = None
-
 # numpy doesn't know bfloat16, define abstract binary type instead
 np_bfloat16 = np.dtype('V2', metadata={"dtype": "bfloat16"})
 np_float8 = np.dtype('V1', metadata={"dtype": "float8"})
@@ -1212,16 +1207,3 @@ def torch_pybind11_abi() -> str:
     if TORCH_PYBIND11_ABI is None:
         TORCH_PYBIND11_ABI = f"{torch._C._PYBIND11_COMPILER_TYPE}{torch._C._PYBIND11_STDLIB}{torch._C._PYBIND11_BUILD_ABI}"
     return TORCH_PYBIND11_ABI
-
-
-@contextmanager
-def unwrap_ray_errors():
-    if ray is None:
-        raise RuntimeError(
-            "Ray requested (TLLM_DISABLE_MPI=1), but not installed. Please install Ray."
-        )
-
-    try:
-        yield
-    except ray.exceptions.RayTaskError as e:
-        raise e.as_instanceof_cause() from e
